+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter } from "react-router-dom";
 
-import RenderRoutes from "./routes";
+import RenderRoutes from "./routes/RenderRoutes";
 
 function App() {
   return (
